backend/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
markdown
django-cors-headers
pytest
pytest-django
pytest-cov
django-extensions
ipdb
djangorestframework-simplejwt